+++ spec/acceptance/example_spec.rb
@@ -25,19 +25,25 @@ module Resync
 
         source_desc = client.get_and_parse(sourcedesc_uri)
         expect(source_desc).to be_a(SourceDescription)
+        expect(source_desc.client).to be(client)
         cap_list_resource = source_desc.resource_for(capability: 'capabilitylist')
+        expect(cap_list_resource.client).to be(client)
         expect(cap_list_resource.uri).to eq(cap_list_uri)
 
         cap_list = cap_list_resource.get_and_parse
         expect(cap_list).to be_a(CapabilityList)
+        expect(cap_list.client).to be(client)
         change_list_resource = cap_list.resource_for(capability: 'changelist')
+        expect(change_list_resource.client).to be(client)
         expect(change_list_resource.uri).to eq(change_list_uri)
 
         change_list = change_list_resource.get_and_parse
+        expect(change_list.client).to be(client)
         expect(change_list).to be_a(ChangeList)
 
         changes = change_list.resources
-        last_update = changes.select { |r| r.metadata.change == Resync::Types::Change::UPDATED }[-1]
+        last_update = changes.select { |r| r.metadata.change == Resync::Types::Change::UPDATED }.to_a[-1]
+        expect(last_update.client).to be(client)
         expect(last_update.uri).to eq(update_uri)
         expect(last_update.get).to eq(update_data)
       end